use BIGINT not NUMERIC, it is faster for arithmetic operations and us…

+++ db/migrations/00004_create_eth_header_cids_table.sql
@@ -5,15 +5,15 @@ CREATE TABLE eth.header_cids (
     parent_hash           VARCHAR(66) NOT NULL,
     cid                   TEXT NOT NULL,
     mh_key                TEXT NOT NULL REFERENCES public.blocks (key) ON DELETE CASCADE DEFERRABLE INITIALLY DEFERRED,
-    td                    NUMERIC NOT NULL,
+    td                    BIGINT NOT NULL,
     node_id               VARCHAR(128) NOT NULL REFERENCES nodes (node_id) ON DELETE CASCADE,
-    reward                NUMERIC NOT NULL,
+    reward                BIGINT NOT NULL,
     state_root            VARCHAR(66) NOT NULL,
     tx_root               VARCHAR(66) NOT NULL,
     receipt_root          VARCHAR(66) NOT NULL,
     uncle_root            VARCHAR(66) NOT NULL,
     bloom                 BYTEA NOT NULL,
-    timestamp             NUMERIC NOT NULL,
+    timestamp             BIGINT NOT NULL,
     times_validated       INTEGER NOT NULL DEFAULT 1,
     base_fee              BIGINT
 );
